feat(build-dotnet): Add conditional output path argument
This commit adds a conditional output path argument to the build step in the ".github/workflows/build-dotnet.yml" file. Now, when building the project(s), the output path will only be specified if the input project file ends with ".sln". This change improves flexibility and allows for more customized build configurations.